refactor(metrics): calculations to separate file
